requirements.txt
streamlit==1.35.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas